requirements.txt
numpy
pandas
matplotlib
seaborn
flask
opencv-python
tensorflow
scikit-learn
pillow
tqdm
pydot
dill
-e .